rename: utils/auth 폴더 추가 및 경로 변경
